Ebs cleanup test (#160)
* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* ebs cleanup test

* allow volume cleanup pipeline

* allow volume cleanup pipeline

* remove apply step

* remove apply step

* remove apply step

* remove apply step

* remove apply step

* remove apply step

* remove apply step

+++ src/ebs_cleanup.sh
@@ -111,7 +111,6 @@ set_date_cmd(){
 
 action() {
   echo $message
-  volumes_found=0
 
   aws_output=$(aws ec2 describe-volumes \
     --region "$region" \
@@ -119,34 +118,33 @@ action() {
     --output text \
     $filters $profile)
 
-  while read -r create_time volume_id state; do
- 
-    volumes_found=$((volumes_found + 1))
-    created_epoch=$($date_cmd -d "$create_time" +%s)
-    age_months_dec=$(awk "BEGIN { printf \"%.1f\", ($now - $created_epoch) / 2592000 }") # 1 decimal place
-    age_months=$(awk "BEGIN { print int($age_months_dec) }")
+  # while read .... do ... <<< $aws_output - this structure because it handles variables better than piping |, e.g. if you wanted to iterate an outside variable within the loop  
+  if [[ -n "$aws_output" ]]; then # because this loop would run once even without any aws_output
+    while read -r create_time volume_id state; do
+      created_epoch=$($date_cmd -d "$create_time" +%s)
+      age_months_dec=$(awk "BEGIN { printf \"%.1f\", ($now - $created_epoch) / 2592000 }") # 1 decimal place
+      age_months=$(awk "BEGIN { print int($age_months_dec) }")
 
-    if [ "$age_months" -ge "$max_age_months" ]; then
-      case $action in
-        all)
-          echo "$volume_id $state $age_months_dec months old in $environment" ;;
-        attached)
-          echo "$volume_id $age_months_dec months oldn $environment" ;;
-        unattached)
-          echo "$volume_id $age_months_dec months old in $environment" ;;
-        delete)
-          if [[ "$dryrun" == true ]]; then
-            echo "Dryrun - would delete volume $volume_id - $age_months_dec months old in $environment"
-          else
-            echo "Deleting volume $volume_id - $age_months_dec months old in $environment"
-            aws ec2 delete-volume --volume-id "$volume_id" --region "$region" $profile
-          fi
-          ;;
-      esac
-    fi
-  done <<< "$aws_output"
-  
-  if [[ "$volumes_found" -eq 0 ]]; then
+      if [ "$age_months" -ge "$max_age_months" ]; then
+        case $action in
+          all)
+            echo "$volume_id $state $age_months_dec months old in $environment" ;;
+          attached)
+            echo "$volume_id $age_months_dec months oldn $environment" ;;
+          unattached)
+            echo "$volume_id $age_months_dec months old in $environment" ;;
+          delete)
+            if [[ "$dryrun" == true ]]; then
+              echo "Dryrun - would delete volume $volume_id - $age_months_dec months old in $environment"
+            else
+              echo "Deleting volume $volume_id - $age_months_dec months old in $environment"
+              aws ec2 delete-volume --volume-id "$volume_id" --region "$region" $profile
+            fi
+            ;;
+        esac
+      fi
+    done <<< "$aws_output"
+  else
     echo $none_message
   fi
 }
